requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
yt-dlp==2024.5.27
pywebpush==2.0.0
aiosqlite==0.20.0
apscheduler==3.10.4
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
cryptography==42.0.8
